# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Isengard MVP API
  version: 0.1.0
  description: REST contract for the M0 plan-run, graph, AI review, and approval workflow.
servers:
  - url: /v1
security:
  - bearerAuth: []
...
paths:
...
  /plan-runs/{planRunId}:
    get:
      tags: [PlanRuns]
      operationId: getPlanRun
      parameters:
        - $ref: "#/components/parameters/PlanRunId"
      responses:
        "200":
          description: Plan run details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanRunEnvelope"
    patch:
      tags: [PlanRuns]
      operationId: updatePlanRun
      parameters:
        - $ref: "#/components/parameters/PlanRunId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlanRunRequest"
      responses:
        "200":
          description: Plan run updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanRunEnvelope"
...
  /plan-runs/{planRunId}/graph:
    get:
      tags: [Graph]
      operationId: getPlanRunGraph
      parameters:
        - $ref: "#/components/parameters/PlanRunId"
      responses:
        "200":
          description: Infrastructure graph DTO for the plan run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphEnvelope"
  /plan-runs/{planRunId}/approvals:
    post:
      tags: [Approvals]
      operationId: requestApplyApproval
      parameters:
        - $ref: "#/components/parameters/PlanRunId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplyApprovalRequest"
      responses:
        "201":
          description: Apply approval requested.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplyApprovalEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PlanRunId:
      name: planRunId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    PlanRunStatus:
      type: string
      enum: [queued, planning, planned, reviewing, needs-approval, approved, applying, applied, failed, cancelled]
    PlanRun:
      type: object
      required: [id, projectId, source, status, cliArgs, createdAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        repositoryId:
          type: string
          format: uuid
        source:
          type: string
          enum: [api, cli, github, scheduler]
        status:
          $ref: "#/components/schemas/PlanRunStatus"
        branch:
          type: string
        commitSha:
          type: string
        terraformVersion:
          type: string
        cliArgs:
          type: array
          items:
            type: string
        summary:
          $ref: "#/components/schemas/TerraformPlanSummary"
        errorMessage:
          type: string
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...
    UpdatePlanRunRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/PlanRunStatus"
        terraformVersion:
          type: string
        summary:
          $ref: "#/components/schemas/TerraformPlanSummary"
        errorMessage:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    TerraformPlanSummary:
      type: object
      required: [add, change, destroy, read, noOp]
      properties:
        add:
          type: integer
          minimum: 0
        change:
          type: integer
          minimum: 0
        destroy:
          type: integer
          minimum: 0
        read:
          type: integer
          minimum: 0
        noOp:
          type: integer
          minimum: 0
...
    Graph:
      type: object
      required: [planRunId, formatVersion, nodes, edges, generatedAt]
      properties:
        planRunId:
          type: string
          format: uuid
        formatVersion:
          type: string
          const: "1.0"
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"
        generatedAt:
          type: string
          format: date-time
    GraphNode:
      type: object
      required: [id, kind, label]
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [project, workspace, module, resource, data-source, provider, output]
        label:
          type: string
        address:
          type: string
        metadata:
          type: object
          additionalProperties: true
    GraphEdge:
      type: object
      required: [id, kind, source, target]
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [contains, depends-on, provides, reads, writes, changes]
        source:
          type: string
        target:
          type: string
        metadata:
          type: object
          additionalProperties: true
    ApplyApproval:
      type: object
      required: [id, planRunId, status, requestedBy, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        planRunId:
          type: string
          format: uuid
        status:
          type: string
          enum: [requested, approved, rejected, expired, cancelled]
        requestedBy:
          type: string
          format: uuid
        approvedBy:
          type: string
          format: uuid
        reason:
          type: string
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateApplyApprovalRequest:
      type: object
      properties:
        reason:
          type: string
        expiresAt:
          type: string
          format: date-time
...
    PlanRunEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelopeBase"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/PlanRun"
...
    GraphEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelopeBase"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/Graph"
    ApplyApprovalEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelopeBase"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/ApplyApproval"
    SuccessEnvelopeBase:
      type: object
      required: [ok, requestId]
      properties:
        ok:
          type: boolean
          const: true
        requestId:
          type: string
...